Made a variety of edits and additions
Parameter sets were added
Functionality to count the total of all open ports was added
The grouping of open ports based on host IP was modified to call a function instead of using a bunch of 'if' statements.  This allows for the code to be more easily edited to add/remove ports of interest.
The calling of the Parse-Nmap.ps1 file was changed to use the invocation operator '&' instead of using 'Invoke-Expression' (a potential security issue).
Other general edits were made along with documentation updates.